Added the context of excersises. Added the simulation from excersise 5, reworked and improved rendering
Classes Simulation and Missile are the solution for new task, these two are really well written I believe. Main got a bit of a rework with split of windows to functions, still no one cares about main me included, it's just a vessel for visualization of excersises. The rendering got reworked, now the center of the window is (0, 0) point. Files got better organized